// src/shared/errors/error-codes.enum.ts
export enum ErrorCodes {
  // Validation errors (400)
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  INVALID_INPUT = 'INVALID_INPUT', // BAD_REQUEST = 'BAD_REQUEST'
  MISSING_REQUIRED_FIELD = 'MISSING_REQUIRED_FIELD',

  // Authentication errors (401)
  UNAUTHORIZED = 'UNAUTHORIZED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',

  // Authorization errors (403)
  FORBIDDEN = 'FORBIDDEN',
  PERMISSION_DENIED = 'PERMISSION_DENIED',

  // Resource errors (404)
  NOT_FOUND = 'NOT_FOUND',

  // Conflict errors (409)
  RESOURCE_ALREADY_EXISTS = 'RESOURCE_ALREADY_EXISTS',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',

  // Business logic errors (422)
  BUSINESS_RULE_VIOLATION = 'BUSINESS_RULE_VIOLATION',
  INVALID_VERIFICATION_CODE = 'INVALID_VERIFICATION_CODE',
  VERIFICATION_CODE_EXPIRED = 'VERIFICATION_CODE_EXPIRED',

  // Server errors (500)
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  DATABASE_ERROR = 'DATABASE_ERROR',
  EXTERNAL_SERVICE_ERROR = 'EXTERNAL_SERVICE_ERROR',
}

// Map error codes to HTTP status codes, if not found, it automatically set to 500
export const HttpStatusCodesMap = new Map<ErrorCodes, number>([
  [ErrorCodes.VALIDATION_FAILED, 400],
  [ErrorCodes.INVALID_INPUT, 400],
  [ErrorCodes.MISSING_REQUIRED_FIELD, 400],

  [ErrorCodes.UNAUTHORIZED, 401],
  [ErrorCodes.INVALID_TOKEN, 401],
  [ErrorCodes.TOKEN_EXPIRED, 401],

  [ErrorCodes.FORBIDDEN, 403],
  [ErrorCodes.PERMISSION_DENIED, 403],

  [ErrorCodes.NOT_FOUND, 404],

  [ErrorCodes.RESOURCE_ALREADY_EXISTS, 409],
  [ErrorCodes.EMAIL_ALREADY_EXISTS, 409],

  [ErrorCodes.BUSINESS_RULE_VIOLATION, 422],
  [ErrorCodes.INVALID_VERIFICATION_CODE, 422],
  [ErrorCodes.VERIFICATION_CODE_EXPIRED, 422],

  [ErrorCodes.INTERNAL_SERVER_ERROR, 500],
  [ErrorCodes.DATABASE_ERROR, 500],
  [ErrorCodes.EXTERNAL_SERVICE_ERROR, 500],
]);